Refactored: Moved entity from the security package to the default entity package
