+++ backend/src/main/java/br/balladesh/pjcappbackend/entity/ArtistEntity.java
@@ -1,6 +1,5 @@
 package br.balladesh.pjcappbackend.entity;
 
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
 import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonValue;
