+++ backend/src/test/java/br/balladesh/pjcappbackend/config/security/services/MyUserDetailsServiceTest.java
@@ -1,6 +1,6 @@
 package br.balladesh.pjcappbackend.config.security.services;
 
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
+import br.balladesh.pjcappbackend.entity.UserEntity;
 import br.balladesh.pjcappbackend.repository.security.UserRepository;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.BeforeEach;
