+++ backend/src/test/java/br/balladesh/pjcappbackend/controllers/api/albums/DeleteAlbumControllerTest.java
@@ -4,7 +4,7 @@ import br.balladesh.pjcappbackend.config.minio.MinIOEndpoint;
 import br.balladesh.pjcappbackend.dto.MessageResponse;
 import br.balladesh.pjcappbackend.entity.AlbumEntity;
 import br.balladesh.pjcappbackend.entity.ArtistEntity;
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
+import br.balladesh.pjcappbackend.entity.UserEntity;
 import br.balladesh.pjcappbackend.repository.AlbumRepository;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Test;
