+++ backend/src/test/java/br/balladesh/pjcappbackend/controllers/api/artists/PutEditArtistControllerTest.java
@@ -2,7 +2,7 @@ package br.balladesh.pjcappbackend.controllers.api.artists;
 
 import br.balladesh.pjcappbackend.dto.api.artists.PutArtistRequestDTO;
 import br.balladesh.pjcappbackend.entity.ArtistEntity;
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
+import br.balladesh.pjcappbackend.entity.UserEntity;
 import br.balladesh.pjcappbackend.repository.ArtistRepository;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.Test;
