+++ backend/src/test/java/br/balladesh/pjcappbackend/dto/api/artists/PagedArtistResponseBodyTest.java
@@ -1,7 +1,7 @@
 package br.balladesh.pjcappbackend.dto.api.artists;
 
 import br.balladesh.pjcappbackend.entity.ArtistEntity;
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
+import br.balladesh.pjcappbackend.entity.UserEntity;
 import com.google.common.collect.ImmutableList;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Test;
