+++ backend/src/test/java/br/balladesh/pjcappbackend/entity/AlbumEntityTest.java
@@ -1,6 +1,5 @@
 package br.balladesh.pjcappbackend.entity;
 
-import br.balladesh.pjcappbackend.entity.security.UserEntity;
 import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
